Add template system
