+++ rescuegroups-sync/templates/block-adoptable-pets.php
@@ -1,0 +1,5 @@
+<ul class="adoptable-pets-block">
+<?php if ( $query->have_posts() ) : while ( $query->have_posts() ) : $query->the_post(); ?>
+    <li><a href="<?php echo esc_url( get_permalink() ); ?>"><?php echo esc_html( get_the_title() ); ?></a></li>
+<?php endwhile; endif; ?>
+</ul>
