+++ rescuegroups-sync/templates/metabox-flags.php
@@ -1,0 +1,4 @@
+<p>
+    <label><input type="checkbox" name="_rescue_sync_featured" value="1" <?php checked( $featured, true ); ?>> <?php echo esc_html__( 'Featured', 'rescuegroups-sync' ); ?></label><br />
+    <label><input type="checkbox" name="_rescue_sync_hidden"   value="1" <?php checked( $hidden, true ); ?>> <?php echo esc_html__( 'Hidden', 'rescuegroups-sync' ); ?></label>
+</p>
